projection logic bugfix

+++ model.h
@@ -1,6 +1,5 @@
 #ifndef __MODEL_H__
 #define __MODEL_H__
-
 #include <vector>
 #include <string>
 #include "geometry.h"
@@ -22,13 +21,13 @@ public:
     int nverts();
     int nfaces();
     Vec3f normal(int iface, int nthvert);
-    Vec3f normal(Vec2i uv);
+    Vec3f normal(Vec2f uv);
     Vec3f vert(int i);
     Vec3f vert(int iface, int nthvert);
-    Vec2i uv(int iface, int nthvert);
-    TGAColor diffuse(Vec2i uv);
-    float specular(Vec2i uv);
+    Vec2f uv(int iface, int nthvert);
+    TGAColor diffuse(Vec2f uv);
+    float specular(Vec2f uv);
     std::vector<int> face(int idx);
 };
-
 #endif //__MODEL_H__
+
